Adding Setup Files :ballot_box_with_check:
Hello everyone ,
Tonight I'm adding new files for setup the project with everything that the project require to work properly.

That's my first Setup.py python file , im testing it ^^.

If it's doesn't work as expected , just follow the current Readme.md available on the Main page of the project on Github.com.

Thank You ,  Francky :+1: